Moved histogram to Celebration station.
