+++ src/components/pages/LandingPage/LoginForm/LoginForm.tsx
@@ -6,7 +6,7 @@ import { token } from '../../../../utils';
 import { useForm, SubmitHandler } from 'react-hook-form';
 import { Input } from '../../../common';
 import { DragonBoi } from '../DragonBoi';
-import { ThoughtBubble } from '../../../common/ThoughtBubble';
+import { ThoughtBubble } from '../../../common';
 
 import welcomeBack from '../../../../assets/welcome-back.png';
