+++ src/styles/index.scss
@@ -22,3 +22,4 @@
 @import './components/common/input.scss';
 @import './components/common/comingSoon.scss';
 @import './components/common/infoHoverTip.scss';
+@import './components/common/histogram.scss';
